third_party/WebKit/LayoutTests/fast/dom/shadow/shadowdom-for-form-associated-element-useragent-expected.txt

CONSOLE WARNING: <shadow> doesn't work for KEYGEN element host.
CONSOLE WARNING: <shadow> doesn't work for FIELDSET element host.
When a OpenShadowRoot is added, ClosedShadowRoot should exists as the oldest Shadow Root.

PASS oldestShadowRoot is not shadowRoot
PASS youngerShadowRoot is shadowRoot
PASS oldestShadowRoot is not shadowRoot
PASS youngerShadowRoot is shadowRoot
PASS oldestShadowRoot is not shadowRoot
PASS youngerShadowRoot is shadowRoot
PASS successfullyParsed is true

TEST COMPLETE
